bandito/requirements.txt
pika
requests
beautifulsoup4
boto3
psycopg2-binary
python-dotenv   # If using .env files for credentials